requirements.txt
asgiref==3.10.0
Django==4.2.16
psycopg2-binary==2.9.11
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
djangorestframework==3.14.0
docker==7.0.0
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
channels==4.0.0
channels-redis==4.1.0
